deprecate old generator
